# metal: more PP speedup for Falcon