--- a/ggml-metal.metal
+++ b/ggml-metal.metal
@@ -610,9 +610,68 @@ kernel void kernel_mul_mat_f16_f32_1row(
 
 }
 
-#define N_F16_F32 4
-
+#define N_F16_F32 8
+#
 kernel void kernel_mul_mat_f16_f32(
+        device const  char * src0,
+        device const  char * src1,
+        device       float * dst,
+        constant   int64_t & ne00,
+        constant   int64_t & ne01,
+        constant   int64_t & ne02,
+        constant  uint64_t & nb00,
+        constant  uint64_t & nb01,
+        constant  uint64_t & nb02,
+        constant   int64_t & ne10,
+        constant   int64_t & ne11,
+        constant   int64_t & ne12,
+        constant  uint64_t & nb10,
+        constant  uint64_t & nb11,
+        constant  uint64_t & nb12,
+        constant   int64_t & ne0,
+        constant   int64_t & ne1,
+        uint3 tgpig[[threadgroup_position_in_grid]],
+        uint3 ntptg[[threads_per_threadgroup]],
+        uint tiitg[[thread_index_in_threadgroup]]) {
+
+    // :MTLSizeMake(ne01, ne11, ne12)
+    const int64_t r0 = tgpig.x * ntptg.x + tiitg;
+    if (r0 >= ne0) {
+        return;
+    }
+    const int64_t r1 = tgpig.y;
+    const int64_t im = tgpig.z;
+
+    //const int64_t im = tiitg/(ne10*ne11);
+    //const int64_t r1 = (tiitg - im*ne10*ne11)/ne10;
+    //const int64_t r0 = tiitg - im*ne10*ne11 - r1*ne10;
+
+    device const half  * x = (device const half  *) (src0 + r0*nb01 + im/(ne12/ne02)*nb02);
+    device const float * y = (device const float *) (src1 + r1*nb11 + im*nb12);
+
+    if (ne00 < 16) {
+        float sumf = 0;
+        for (int i = 0; i < ne00; ++i) {
+            sumf += (float) x[i] * (float) y[i];
+        }
+        dst[im*ne1*ne0 + r1*ne0 + r0] = sumf;
+    }
+    else {
+        float sumf = 0;
+        device const half4  * x4 = (device const half4  *) x;
+        device const float4 * y4 = (device const float4 *) y;
+        for (int i = 0; i < ne00/4; ++i) {
+            for (int k = 0; k < 4; ++k) sumf += (float) x4[i][k] * y4[i][k];
+        }
+        for (int i = 4*(ne00/4); i < ne00; ++i) {
+            sumf += (float) x[i] * y[i];
+        }
+        dst[im*ne1*ne0 + r1*ne0 + r0] = sumf;
+
+    }
+}
+
+kernel void kernel_mul_mat_f16_f32_old(
         device const  char * src0,
         device const  char * src1,
         device       float * dst,
@@ -639,7 +698,7 @@ kernel void kernel_mul_mat_f16_f32(
 
     device const half * x = (device const half *) (src0 + r0*nb01 + im/(ne12/ne02)*nb02);
 
-    if (ne00 < 128) {
+    if (ne00 < 64) { //128) {
         for (int row = 0; row < N_F16_F32; ++row) {
             int r1 = rb + row;
             if (r1 >= ne11) {
@@ -659,27 +718,47 @@ kernel void kernel_mul_mat_f16_f32(
             }
         }
     } else {
+        const int ix = tiisg/N_F16_F32;
+        const int iy = tiisg%N_F16_F32;
+        const int r1 = rb + iy < ne11 ? rb + iy : ne11-1;
+        float sumf[N_F16_F32] = {0.f};
         device const half4 * x4 = (device const half4 *)x;
+        device const float  * y  = (device const float  *) (src1 + r1*nb11 + im*nb12);
+        device const float4 * y4 = (device const float4 *) y;
+        for (int i = ix; i < ne00/4; i += 32/N_F16_F32) {
+            for (int k = 0; k < 4; ++k) sumf[iy] += (float) x4[i][k] * y4[i][k];
+        }
+        for (int i = 4*(ne00/4)+ix; i < ne00; i += 32/N_F16_F32) {
+            sumf[iy] += (float) x[i] * y[i];
+        }
         for (int row = 0; row < N_F16_F32; ++row) {
-            int r1 = rb + row;
-            if (r1 >= ne11) {
-                break;
-            }
-
-            device const float  * y  = (device const float  *) (src1 + r1*nb11 + im*nb12);
-            device const float4 * y4 = (device const float4 *) y;
-
-            float sumf = 0;
-            for (int i = tiisg; i < ne00/4; i += 32) {
-                for (int k = 0; k < 4; ++k) sumf += (float) x4[i][k] * y4[i][k];
-            }
-
-            float all_sum = simd_sum(sumf);
-            if (tiisg == 0) {
-                for (int i = 4*(ne00/4); i < ne00; ++i) all_sum += (float) x[i] * y[i];
-                dst[im*ne1*ne0 + r1*ne0 + r0] = all_sum;
+            float all_sum = simd_sum(sumf[row]);
+            if (tiisg == 0 && rb + row < ne11) {
+                dst[im*ne1*ne0 + (rb + row)*ne0 + r0] = all_sum;
             }
         }
+
+        //device const half4 * x4 = (device const half4 *)x;
+        //for (int row = 0; row < N_F16_F32; ++row) {
+        //    int r1 = rb + row;
+        //    if (r1 >= ne11) {
+        //        break;
+        //    }
+
+        //    device const float  * y  = (device const float  *) (src1 + r1*nb11 + im*nb12);
+        //    device const float4 * y4 = (device const float4 *) y;
+
+        //    float sumf = 0;
+        //    for (int i = tiisg; i < ne00/4; i += 32) {
+        //        for (int k = 0; k < 4; ++k) sumf += (float) x4[i][k] * y4[i][k];
+        //    }
+
+        //    float all_sum = simd_sum(sumf);
+        //    if (tiisg == 0) {
+        //        for (int i = 4*(ne00/4); i < ne00; ++i) all_sum += (float) x[i] * y[i];
+        //        dst[im*ne1*ne0 + r1*ne0 + r0] = all_sum;
+        //    }
+        //}
     }
 }