… examples/llama-bench/llama-bench.cpp