@ikawrakow 2023-09-14
Sure, copying `Q` to contiguous memory is an obvious approach, but I wanted to avoid the copy to not have the TG performance degradation. For LLaMA-7B the `Q` tensor is 128 x Nb x 32, so the amount of data to be copied becomes non-trivial when generating a long sequence of tokens.