@ggerganov 2023-09-14
During TG, `Nb == 1`, so the `Q` tensor is always the same size and it is small. Here is for example the graph for `n_past == 45` and `n_batch == 1`:

![image](https://github.com/ggerganov/llama.cpp/assets/1991296/93ac4a44-e7e6-4a9e-a332-b38c642847cc)